Move `eval_jsonpath` and `build_safe_lambda` to expressions/common.py
